Refactor out some common GAN things
