+++ satflow/models/gan/__init__.py
@@ -1,0 +1,1 @@
+from .discriminators import GANLoss, PixelDiscriminator, NLayerDiscriminator
